refactor code for categorization purposes
