+++ helpers/http.js
@@ -1,4 +1,5 @@
 const axios = require('axios');
+const { getSymbolFromColor, touchBarArray } = require('./touchbar');
 
 /**
  * Returns coinbase http request based on url argument
@@ -15,6 +16,20 @@ const getCurrencyData = async (url) => {
     })
 }
 
+/**
+ * Fetches coinbase currency data and updates object argument label
+ */
+const updateCurrencyData = async () => {
+  touchBarArray.forEach(async ({_label, _backgroundColor}, index) => {
+    if(_label) {
+      let currencySymbol = getSymbolFromColor(_backgroundColor); // get currency symbol
+      const { data } = await getCurrencyData(`https://api.coinbase.com/v2/prices/${currencySymbol}-USD/spot`); // make request using currency symbol
+      touchBarArray[index].label = "$" + data.amount; // update touchbar array object label with fetched amount
+    }
+  })
+}
+
 module.exports = {
-  getCurrencyData
+  getCurrencyData,
+  updateCurrencyData
 }
